+++ dgraph/Makefile
@@ -19,6 +19,7 @@ BUILD         ?= $(shell git rev-parse --short HEAD)
 BUILD_DATE    ?= $(shell git log -1 --format=%ci)
 BUILD_BRANCH  ?= $(shell git rev-parse --abbrev-ref HEAD)
 BUILD_VERSION ?= $(shell git describe --always --tags)
+BUILD_TAGS    ?=
 
 # Build-time Go variables
 dgraphVersion  = github.com/dgraph-io/dgraph/x.dgraphVersion
@@ -26,13 +27,19 @@ gitBranch      = github.com/dgraph-io/dgraph/x.gitBranch
 lastCommitSHA  = github.com/dgraph-io/dgraph/x.lastCommitSHA
 lastCommitTime = github.com/dgraph-io/dgraph/x.lastCommitTime
 
-BUILD_FLAGS   ?= -ldflags '-X ${lastCommitSHA}=${BUILD} -X "${lastCommitTime}=${BUILD_DATE}" -X ${dgraphVersion}=${BUILD_VERSION} -X ${gitBranch}=${BUILD_BRANCH}'
+BUILD_FLAGS   ?= -ldflags '-X ${lastCommitSHA}=${BUILD} -X "${lastCommitTime}=${BUILD_DATE}" -X "${dgraphVersion}=${BUILD_VERSION}" -X ${gitBranch}=${BUILD_BRANCH}'
+
+# Insert build tags if specified
+ifneq ($(strip $(BUILD_TAGS)),)
+	BUILD_FLAGS += -tags '$(BUILD_TAGS)'
+	BUILD_VERSION := $(BUILD_VERSION)_oss
+endif
 
 .PHONY: all $(BIN)
 all: $(BIN)
 
 $(BIN): clean
-	@go build $(BUILD_FLAGS) -o $(BIN)
+	go build $(BUILD_FLAGS) -o $(BIN)
 
 clean:
 	@rm -f $(BIN)
